# library prj/python-flask-server-generated/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Library API - Full Specification
  description: "API Thư viện sách, bao gồm phân trang, Reviews và bảo mật JWT."
  version: 1.0.0
servers:
- url: http://127.0.0.1:5000
paths:
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Đăng nhập và nhận JWT Token
      operationId: api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
        required: true
      responses:
        "200":
          description: "Đăng nhập thành công, trả về JWT Token."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Lỗi xác thực (Thiếu/Sai/Hết hạn Token).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Lỗi máy chủ không xác định.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-openapi-router-controller: swagger_server.controllers.authentication_controller
  /api/v1/books:
    get:
      tags:
      - Books
      summary: Lấy danh sách sách có phân trang Offset/Limit
      operationId: api_v1_books_get
      parameters:
      - name: offset
        in: query
        description: Vị trí bắt đầu của danh sách (số lượng bản ghi bỏ qua). Mặc định
          0.
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: limit
        in: query
        description: Số lượng bản ghi tối đa được trả về. Mặc định 10.
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 10
      responses:
        "200":
          description: "Danh sách sách được phân trang, trả về cùng metadata."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedBookList"
        "400":
          description: Lỗi dữ liệu đầu vào.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Lỗi máy chủ không xác định.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-openapi-router-controller: swagger_server.controllers.books_controller
    post:
      tags:
      - Books
      summary: Thêm sách mới (Yêu cầu Admin)
      operationId: api_v1_books_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
        required: true
      responses:
        "201":
          description: Sách được tạo thành công.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
        "400":
          description: Lỗi dữ liệu đầu vào.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Lỗi xác thực (Thiếu/Sai/Hết hạn Token).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.books_controller
  /api/v1/books/{book_id}:
    get:
      tags:
      - Books
      summary: Lấy chi tiết sách theo ID
      operationId: api_v1_books_book_id_get
      parameters:
      - name: book_id
        in: path
        description: ID của sách
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Thông tin chi tiết của sách.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Không tìm thấy tài nguyên.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Lỗi máy chủ không xác định.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-openapi-router-controller: swagger_server.controllers.books_controller
    put:
      tags:
      - Books
      summary: Cập nhật sách (Yêu cầu Admin)
      operationId: api_v1_books_book_id_put
      parameters:
      - name: book_id
        in: path
        description: ID của sách
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
        required: true
      responses:
        "200":
          description: Cập nhật sách thành công.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
        "401":
          description: Lỗi xác thực (Thiếu/Sai/Hết hạn Token).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Lỗi cấp quyền.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Không tìm thấy tài nguyên.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.books_controller
    delete:
      tags:
      - Books
      summary: Xóa sách (Yêu cầu Admin)
      operationId: api_v1_books_book_id_delete
      parameters:
      - name: book_id
        in: path
        description: ID của sách
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "204":
          description: Xóa sách thành công (Không trả về nội dung).
        "401":
          description: Lỗi xác thực (Thiếu/Sai/Hết hạn Token).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Lỗi cấp quyền.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Không tìm thấy tài nguyên.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.books_controller
  /api/v1/books/{book_id}/reviews:
    get:
      tags:
      - Reviews
      summary: Lấy danh sách nhận xét của một cuốn sách (có phân trang)
      operationId: api_v1_books_book_id_reviews_get
      parameters:
      - name: book_id
        in: path
        description: ID của sách
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: offset
        in: query
        description: Vị trí bắt đầu của danh sách (số lượng bản ghi bỏ qua). Mặc định
          0.
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: limit
        in: query
        description: Số lượng bản ghi tối đa được trả về. Mặc định 10.
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 10
      responses:
        "200":
          description: Danh sách nhận xét được phân trang.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedReviewList"
        "404":
          description: Không tìm thấy tài nguyên.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-openapi-router-controller: swagger_server.controllers.reviews_controller
    post:
      tags:
      - Reviews
      summary: Thêm nhận xét mới cho sách (Yêu cầu Đăng nhập)
      operationId: api_v1_books_book_id_reviews_post
      parameters:
      - name: book_id
        in: path
        description: ID của sách
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
        required: true
      responses:
        "201":
          description: Nhận xét được tạo thành công.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponse"
        "400":
          description: Lỗi dữ liệu đầu vào.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Lỗi xác thực (Thiếu/Sai/Hết hạn Token).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Không tìm thấy tài nguyên.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - BearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.reviews_controller
  /api/v1/books/code/reading-time:
    get:
      tags:
      - Utility
      summary: Cung cấp mã JavaScript tính thời gian đọc (Code-On-Demand)
      operationId: api_v1_books_code_reading_time_get
      responses:
        "200":
          description: Trả về mã JavaScript để client thực thi.
          content:
            application/javascript:
              schema:
                type: string
                format: binary
                example: !!binary |-
                  ZnVuY3Rpb24gY2FsY3VsYXRlUmVhZGluZ1RpbWUod29yZENvdW50KSB7IC8qIC4uLiAqLyB9
                x-content-type: application/javascript
        default:
          description: Lỗi máy chủ không xác định.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-openapi-router-controller: swagger_server.controllers.utility_controller
components:
  schemas:
    LoginInput:
      required:
      - password
      - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
        token_type:
          type: string
      example:
        message: message
        token_type: token_type
        token: token
    Book:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
      example:
        author: author
        title: title
    BookInput:
      required:
      - author
      - title
      type: object
      properties:
        title:
          type: string
        author:
          type: string
    BookResponse:
      type: object
      properties:
        message:
          type: string
        book_id:
          type: integer
        book:
          $ref: "#/components/schemas/Book"
      example:
        book:
          author: author
          title: title
        book_id: 0
        message: message
    Review:
      type: object
      properties:
        book_id:
          type: integer
          description: ID của sách được đánh giá
        user:
          type: string
          description: Vai trò người dùng đã đánh giá
        rating:
          maximum: 5
          minimum: 1
          type: integer
        comment:
          type: string
      example:
        rating: 3
        comment: comment
        book_id: 0
        user: user
    ReviewInput:
      required:
      - rating
      type: object
      properties:
        rating:
          maximum: 5
          minimum: 1
          type: integer
        comment:
          type: string
    ReviewResponse:
      type: object
      properties:
        message:
          type: string
        review_id:
          type: integer
        review:
          $ref: "#/components/schemas/Review"
      example:
        review_id: 0
        review:
          rating: 3
          comment: comment
          book_id: 0
          user: user
        message: message
    PaginationMetadata:
      type: object
      properties:
        total_count:
          type: integer
        current_items:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        next_offset:
          type: integer
          nullable: true
      description: Metadata về trạng thái phân trang.
      example:
        offset: 1
        total_count: 0
        limit: 5
        current_items: 6
        next_offset: 5
    PagedBookList:
      type: object
      properties:
        data:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Book"
          description: Tập hợp các đối tượng sách (key là ID sách).
        pagination:
          $ref: "#/components/schemas/PaginationMetadata"
      description: Cấu trúc phản hồi cho danh sách sách có phân trang.
      example:
        pagination:
          offset: 1
          total_count: 0
          limit: 5
          current_items: 6
          next_offset: 5
        data:
          key:
            author: author
            title: title
    PagedReviewList:
      type: object
      properties:
        data:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Review"
          description: Tập hợp các đối tượng nhận xét (key là ID nhận xét).
        pagination:
          $ref: "#/components/schemas/PaginationMetadata"
      description: Cấu trúc phản hồi cho danh sách nhận xét có phân trang.
      example:
        pagination:
          offset: 1
          total_count: 0
          limit: 5
          current_items: 6
          next_offset: 5
        data:
          key:
            rating: 3
            comment: comment
            book_id: 0
            user: user
    Error:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequestError:
      description: Lỗi dữ liệu đầu vào.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFoundError:
      description: Không tìm thấy tài nguyên.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnauthorizedError:
      description: Lỗi xác thực (Thiếu/Sai/Hết hạn Token).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ForbiddenError:
      description: Lỗi cấp quyền.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: Lỗi máy chủ không xác định.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  parameters:
    BookIdParam:
      name: book_id
      in: path
      description: ID của sách
      required: true
      style: simple
      explode: false
      schema:
        type: integer
    OffsetParam:
      name: offset
      in: query
      description: Vị trí bắt đầu của danh sách (số lượng bản ghi bỏ qua). Mặc định
        0.
      required: false
      style: form
      explode: true
      schema:
        minimum: 0
        type: integer
        default: 0
    LimitParam:
      name: limit
      in: query
      description: Số lượng bản ghi tối đa được trả về. Mặc định 10.
      required: false
      style: form
      explode: true
      schema:
        maximum: 100
        minimum: 1
        type: integer
        default: 10
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_BearerAuth